requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.0.12
decorator==5.1.1
Django==4.0.5
django-filter==22.1
djangorestframework==3.13.1
idna==3.3
importlib-metadata==4.11.4
Markdown==3.3.7
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
urllib3==1.26.9
validators==0.20.0
zipp==3.8.0
